# src/swagger-docs/store.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Store",
        "description": "",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
...
    "host": "localhost:30301",
    "basePath": "/api/v1/stores",
    "paths": {
...
        "/{id}/customers": {
            "get": {
                "tags": [
                    "Get Customers for given Store Id"
                ],
                "description": "Get Customers for given Store Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
...
}